TX ONLY working... RX not working for some reason...potentially the board?

+++ STM32/stm32_workspace/spudglo/src/bsp/flash_info.h
@@ -12,14 +12,14 @@
 #define FLASH_MAX_PAGE_NUMBER                       ((128 * 2048 / FLASH_PAGE_SIZE)) - 1
 #define FLASH_MIN_PAGE_NUMBER                       0
 
-#define FLASH_USER_SPACE_START_ADDRESS              ((uint64_t)250 * 2048)
+#define FLASH_USER_SPACE_START_ADDRESS              ((uint32_t)250 * 2048)
 
-#define FLASH_START_ADDRESS                         ((uint64_t)0x08000000)
-#define FLASH_STOP_ADDRESS                         	((uint64_t)0x08020000)
+#define FLASH_START_ADDRESS                         ((uint32_t)0x08000000)
+#define FLASH_STOP_ADDRESS                         	((uint32_t)0x08020000)
 
-#define FLASH_SUB_BLOCK_CONFIG_ADDRESS_OFFSET       ((uint64_t)(FLASH_MAX_PAGE_NUMBER - 3) * 2048)
-//#define FLASH_SUB_BLOCK_STATUS_ADDRESS_OFFSET   ((uint64_t)251 * 2048)
-//#define FLASH_SUB_BLOCK_IMAGE_ADDRESS_OFFSET    ((uint64_t)252 * 2048)
+#define FLASH_SUB_BLOCK_CONFIG_ADDRESS_OFFSET       ((uint32_t)(FLASH_MAX_PAGE_NUMBER - 3) * 2048)
+//#define FLASH_SUB_BLOCK_STATUS_ADDRESS_OFFSET   ((uint32_t)251 * 2048)
+//#define FLASH_SUB_BLOCK_IMAGE_ADDRESS_OFFSET    ((uint32_t)252 * 2048)
 
 #define FLASH_INFO_SUB_BLOCK_SECTOR_SIZE_BYTES      (2048)
